OMAP: tidspbridge: add assert/deassert reset functions and use them instead of directly accessing PRM registers
