+++ include/linux/platform_data/dsp-omap.h
@@ -2,6 +2,7 @@
 #define __OMAP_DSP_H__
 
 #include <linux/types.h>
+#include <linux/platform_device.h>
 
 struct omap_dsp_platform_data {
 	void (*dsp_set_min_opp) (u8 opp_id);
@@ -20,6 +21,9 @@ struct omap_dsp_platform_data {
 
 	void (*set_bootaddr)(u32);
 	void (*set_bootmode)(u8);
+
+	int (*assert_reset)(struct platform_device *pdev);
+	int (*deassert_reset)(struct platform_device *pdev);
 };
 
 #endif
